requirements.txt
# Backend
fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.7.0
aiohttp>=3.9.5
python-dotenv>=1.0.1
cachetools>=5.3.3

# ML
prophet>=1.1.5
scikit-learn>=1.4.0
pandas>=2.2.0
numpy>=1.26.0
joblib>=1.4.0

# Evaluación y gráficas
matplotlib>=3.8.0

# Utils
httpx>=0.27.0
requests>=2.28.0
